refactor(cli): split self_cmd.rs into a self_cmd/ module directory
self_cmd.rs (2415 lines) becomes self_cmd/{version,update,download,migrate}.rs
plus a small exit.rs (exit-code consts), tests mirrored under self_cmd/tests/.
version = identity + check-for-update; update = orchestration + defer/apply
machinery; download = binary fetch/verify/swap; migrate = cartog self
migrate-db. mod.rs re-exports the 5 public names unchanged. Test-only
#[cfg(test)] peer-lock helpers stay with prod. Cross-module helpers bumped to
pub(crate). No prod logic or test changed (51 tests, 5 #[serial] preserved).
