+++ crates/cartog/src/commands/self_cmd/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Unit tests for the `cartog self` command group, split by concern.
+
+mod download;
+mod migrate;
+mod update;
+mod version;
